move: DefaultCommandTemplate to inner class of CommandTemplate
