+++ src/main/java/lanat/CommandTemplate.java
@@ -92,6 +92,7 @@ import java.util.List;
  *     <li>{@link CommandTemplate#afterInit(Command)}: Called after the Command is initialized.</li>
  * </ul>
  * <p>
+ * @see CommandTemplate.Default
  */
 @Command.Define
 public abstract class CommandTemplate {
@@ -181,5 +182,37 @@ public abstract class CommandTemplate {
 		return annotationNames.length == 0 ?
 			new String[] { cmdTemplate.getSimpleName() }
 			: annotationNames;
+	}
+
+	/**
+	 * A default command template that adds the 'help' and 'version' arguments to the command.
+	 * <ul>
+	 * <li>The 'help' argument shows the help message of the command (provided by the {@link Command#getHelp()} method).</li>
+	 * <li>The 'version' argument shows the version of the program (provided by the {@link ArgumentParser#getVersion()} method).</li>
+	 * </ul>
+	 */
+	@Command.Define
+	public static class Default extends CommandTemplate {
+		/*
+		 * The reason we add these arguments here is so that they do not "physically" appear in the
+		 * actual class that extends this one. 'help' and 'version' are just
+		 * arguments that execute actions, and they not really provide any useful values.
+		 */
+		@InitDef
+		public static void afterInit(@NotNull Command cmd) {
+			cmd.addArgument(Argument.createOfBoolType("help")
+				.onOk(t -> System.out.println(cmd.getHelp()))
+				.withDescription("Shows this message.")
+				.allowsUnique()
+			);
+
+			if (cmd instanceof ArgumentParser ap) {
+				cmd.addArgument(Argument.createOfBoolType("version")
+					.onOk(t -> System.out.println("Version: " + ap.getVersion()))
+					.withDescription("Shows the version of this program.")
+					.allowsUnique()
+				);
+			}
+		}
 	}
 }
